refetch user profile data on clientside navigation
