+++ apps/site/src/lib/blocks.ts
@@ -228,18 +228,6 @@ export const readBlocksFromDisk = async (): Promise<
   return result;
 };
 
-// Blocks which are currently not compliant with the spec, and are thus misleading examples
-const blocksToHide = ["@hash/embed"];
-
-/** Helps consistently hide certain blocks from the Hub and user profile pages */
-export const excludeHiddenBlocks = (
-  blocks: ExpandedBlockMetadata[],
-): ExpandedBlockMetadata[] => {
-  return blocks.filter(
-    ({ pathWithNamespace }) => !blocksToHide.includes(pathWithNamespace),
-  );
-};
-
 export const retrieveBlockFileContent = async ({
   pathWithNamespace,
   schema: metadataSchemaUrl,
