+++ apps/site/src/pages/[shortname]/[[...profile-tabs]].page.tsx
@@ -1,6 +1,7 @@
 import { GetStaticPaths, GetStaticProps, NextPage } from "next";
-import Error from "next/error";
+import NextError from "next/error";
 import { useRouter } from "next/router";
+import { useEffect, useRef, useState } from "react";
 
 import { TABS } from "../../components/pages/user/tabs";
 import {
@@ -8,7 +9,7 @@ import {
   UserPageProps,
 } from "../../components/pages/user/user-page-component";
 import { apiClient } from "../../lib/api-client";
-import { excludeHiddenBlocks } from "../../lib/blocks";
+import { excludeHiddenBlocks } from "../../lib/excluded-blocks";
 
 type UserPageQueryParams = {
   "profile-tabs": string[];
@@ -24,6 +25,32 @@ export const getStaticPaths: GetStaticPaths = async () => {
   return {
     paths: [],
     fallback: "blocking",
+  };
+};
+
+const fetchUserProfileData = async (shortname: string) => {
+  const [userResponse, blocksResponse, entityTypesResponse] = await Promise.all(
+    [
+      apiClient.getUser({
+        shortname: shortname.replace("@", ""),
+      }),
+      apiClient.getUserBlocks({
+        shortname: shortname.replace("@", ""),
+      }),
+      apiClient.getUserEntityTypes({
+        shortname: shortname.replace("@", ""),
+      }),
+    ],
+  );
+
+  if (!userResponse.data?.user) {
+    throw new Error("No user found");
+  }
+
+  return {
+    blocks: excludeHiddenBlocks(blocksResponse.data?.blocks || []),
+    entityTypes: entityTypesResponse.data?.entityTypes || [],
+    user: userResponse.data.user,
   };
 };
 
@@ -43,42 +70,50 @@ export const getStaticProps: GetStaticProps<
     };
   }
 
-  const [userResponse, blocksResponse, entityTypesResponse] = await Promise.all(
-    [
-      apiClient.getUser({
-        shortname: shortname.replace("@", ""),
-      }),
-      apiClient.getUserBlocks({
-        shortname: shortname.replace("@", ""),
-      }),
-      apiClient.getUserEntityTypes({
-        shortname: shortname.replace("@", ""),
-      }),
-    ],
-  );
-
-  if (userResponse.error || !userResponse.data) {
+  try {
+    const { blocks, entityTypes, user } = await fetchUserProfileData(shortname);
+    return {
+      props: {
+        blocks,
+        entityTypes,
+        user,
+      },
+      revalidate: 60,
+    };
+  } catch {
     return { notFound: true, revalidate: 60 };
   }
-
-  return {
-    props: {
-      blocks: excludeHiddenBlocks(blocksResponse.data?.blocks || []),
-      entityTypes: entityTypesResponse.data?.entityTypes || [],
-      user: userResponse.data.user,
-    },
-    revalidate: 60,
-  };
 };
 
-const UserPage: NextPage<UserPageProps> = ({ user, blocks, entityTypes }) => {
+const UserPage: NextPage<UserPageProps> = ({
+  user: prerenderedUser,
+  blocks: prerenderedBlocks,
+  entityTypes: prerenderedEntityTypes,
+}) => {
   const router = useRouter();
+  const hasMounted = useRef(false);
 
   const matchingTab = findTab(router.query["profile-tabs"]);
 
+  const [{ user, blocks, entityTypes }, updateUserProfileData] = useState({
+    user: prerenderedUser,
+    blocks: prerenderedBlocks,
+    entityTypes: prerenderedEntityTypes,
+  });
+
+  useEffect(() => {
+    if (hasMounted.current) {
+      void fetchUserProfileData(user.shortname!).then((userProfileData) => {
+        updateUserProfileData(userProfileData);
+      });
+    } else {
+      hasMounted.current = true;
+    }
+  }, [user.shortname]);
+
   // Protect against unlikely client-side navigation to a non-existing profile tab
   if (!matchingTab) {
-    return <Error statusCode={404} />;
+    return <NextError statusCode={404} />;
   }
 
   return (
